requirements.txt
PyQt5
pymodbus
pyserial
